Make tombstoning a cross-device operation
